backend/requirements.txt
# Core requirements for Reynard Backend
# Generated from pyproject.toml for compatibility

# Core dependencies
fastapi
uvicorn[standard]
python-jose[cryptography]
python-dotenv
python-multipart
email-validator
pydantic[email]
slowapi
limits
torch
torchvision
pillow
numpy
sqlalchemy
psycopg2-binary
aiohttp
pgvector
argon2-cffi
cryptography
sentence-transformers
transformers
torch
rank-bm25
numpy
